Program.cs
Uploaded working main code of file. As mentioned before, this is very similar to my ConsoleExperienceCalculator project, but with a few minor differences made, so that a friend of mine could use the program on his computer, with his own specifications. Main differences in the code will show up more clearly in the Character.cs code.